[IMP] fruthos: Print Invoice Reports Buttons
Hiding the Print actions (context) that print the invoice reports
and adding the report "Invoice Receipt" as a button in the form view,
and the other reports are not needed.

Also, adding a new field `is_credit_invoice` in the model `account.invoice`
to know if the invoice is a credit or a cash invoice, to know when to
show the report button, and the "Sent & Print" button. If the invoice is
Cash the buttons will only be available when the invoiced is
completely paid, otherwise, it will be available as soon as it's
confirmed.

And updating the field `is_credit_invoice` because in production already
exists and there are records in NULL.