[IMP] l10n_cr_hr_payroll: Holiday provision
Some companies saved money each payslip period to pay the days that the
employee will enjoy his vacations/holidays. This action we called a
holidays provision.

Now new company contribution rules are added to allow user make this
savings, a new rule for vacations will be used if this option is used,
this new rule will allow the user specify the amount that will be paid
as vacation and defined the correct account accounts configuration.

Task#54116