[FIX] fruthos: Avoid traceback invoice pdf wrong t-options
In a previos commit, the t-field-options were updated by its new form
t-options, but the value was wrong for this new form, and was given a
traceback error when the ticket pdf was trying to be genereted.

Task#54116

[FIX] fruthos: incorrect name of a payroll manual input
The name of an manual input was incorrect, actually the rule has the
correct name but here was typed incorrectly.

Issue#19542